New command: 'matrix' (discrete colormap shortcut where the resolution is automatically adapted to the data)
'matrix "foo.z"' corresponds to 'colormap "foo.z" 0 0 interpolate discrete' and takes the same optional arguments as a standard colormap command. When using the 'matrix' command (or colormap 0 0 interpolate discrete), the number of cells to be drawn is automatically adapted to the dataset and the axis' limits (e.g., for the full range the resolution is exactly nx ny as specified in the 3D data file).